Refactor: follow router 분리
