+++ backend/src/routers/v1/users/follow.ts
@@ -1,0 +1,53 @@
+import { Request, Response } from 'express';
+import { Controller, Req, Res, Get, Post, Delete, UseBefore } from 'routing-controllers';
+import * as passport from 'passport';
+
+import { FollowService } from 'src/services';
+import { ERROR, RESPONSECODE } from 'src/utils';
+
+@Controller('/users')
+export default class UsersRouter {
+  @Get('/:userID/follows')
+  async getUserFollows(@Req() request: Request, @Res() response: Response) {
+    const { userID } = request.params;
+    const follows = await FollowService.findFollows(userID);
+    return response.json({ code: RESPONSECODE.SUCCESS, data: follows });
+  }
+
+  @Get('/:userID/followers')
+  async getUserFollowers(@Req() request: Request, @Res() response: Response) {
+    const { userID } = request.params;
+    const followList = await FollowService.findFollowers(userID);
+    return response.json({ code: RESPONSECODE.SUCCESS, data: followList });
+  }
+
+  @Post('/:userID/follows')
+  @UseBefore(passport.authenticate('jwt-registered', { session: false }))
+  async putUserFollows(@Req() request: Request, @Res() response: Response) {
+    const { userID } = request.params;
+    const { userID: bodyUserID } = request.body;
+    if (bodyUserID !== request.user?.userID) {
+      throw new Error(ERROR.PERMISSION_DENIED);
+    }
+    if (userID === request.user!.userID) {
+      throw new Error(ERROR.WRONG_PARAMS_TYPE);
+    }
+    await FollowService.createFollow(userID, request.user!.userID);
+    return response.json({ code: RESPONSECODE.SUCCESS });
+  }
+
+  @Delete('/:userID/follows')
+  @UseBefore(passport.authenticate('jwt-registered', { session: false }))
+  async deleteUserFollows(@Req() request: Request, @Res() response: Response) {
+    const { userID } = request.params;
+    const { userID: bodyUserID } = request.body;
+    if (bodyUserID !== request.user?.userID) {
+      throw new Error(ERROR.PERMISSION_DENIED);
+    }
+    if (userID === request.user!.userID) {
+      throw new Error(ERROR.WRONG_PARAMS_TYPE);
+    }
+    await FollowService.removeFollow(request.user!.userID, userID);
+    return response.json({ code: RESPONSECODE.SUCCESS });
+  }
+}
